+++ Assets/Prefabs/Boss.prefab
@@ -1268,7 +1268,9 @@ GameObject:
   - component: {fileID: 6679307635143848697}
   - component: {fileID: 5295105000875886994}
   - component: {fileID: 6950495233389089913}
-  - component: {fileID: 1179444589407780789}
+  - component: {fileID: 786958631762411733}
+  - component: {fileID: 661711793134926653}
+  - component: {fileID: 8928569220642699476}
   - component: {fileID: 7033277526492962462}
   m_Layer: 0
   m_Name: Boss
@@ -1322,7 +1324,7 @@ Rigidbody:
   m_Interpolate: 2
   m_Constraints: 80
   m_CollisionDetection: 0
---- !u!114 &1179444589407780789
+--- !u!114 &786958631762411733
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -1331,15 +1333,41 @@ MonoBehaviour:
   m_GameObject: {fileID: 8407359795339301912}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5a90cfbb14448f24b85110c1453bc843, type: 3}
+  m_Script: {fileID: 11500000, guid: 13b59b0c6d24e174ca3eb084293b6d4c, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  bombPref: {fileID: 4151298751684108331, guid: dfa936be66d6ad043bf72479c5139eb3,
+    type: 3}
+  bombStartPos: {fileID: 8976917753214957289}
+  force: 0
+--- !u!114 &661711793134926653
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8407359795339301912}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 00a9fb7e80fe93f42b426b2846c2f8db, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  healthLine: {fileID: 0}
+--- !u!114 &8928569220642699476
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8407359795339301912}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 3096a70c92a023e4788f380e176df0a7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   visionRadius: 0
   attackDistance: 0
   playerPos: {fileID: 0}
-  bombPref: {fileID: 0}
-  bombStartPos: {fileID: 0}
-  force: 0
 --- !u!95 &7033277526492962462
 Animator:
   serializedVersion: 3
