+++ Assets/Prefabs/Enemy.prefab
@@ -852,7 +852,9 @@ GameObject:
   - component: {fileID: 3457622948174429050}
   - component: {fileID: 2228399938504628195}
   - component: {fileID: 741127585833357989}
-  - component: {fileID: -7138211938627965860}
+  - component: {fileID: 8392190425227341755}
+  - component: {fileID: 1530201395946127550}
+  - component: {fileID: 52368431707081983}
   - component: {fileID: 349551690098673573}
   m_Layer: 0
   m_Name: Enemy
@@ -928,7 +930,7 @@ NavMeshAgent:
   m_BaseOffset: 0
   m_WalkableMask: 4294967295
   m_ObstacleAvoidanceType: 4
---- !u!114 &-7138211938627965860
+--- !u!114 &8392190425227341755
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -937,14 +939,38 @@ MonoBehaviour:
   m_GameObject: {fileID: 9110048099626921029}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 6f732432a9accaa49b8a6b130de56817, type: 3}
+  m_Script: {fileID: 11500000, guid: 747355fd32eb1d246883d5e93c895d92, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   visionRadius: 8
   attackDistance: 4
+--- !u!114 &1530201395946127550
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9110048099626921029}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9559730328ae28b458093c7ad98d61a3, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   bulletPref: {fileID: 5977638688046338175, guid: 78741e33b87fc8d40b825bc3b601f8e6,
     type: 3}
   bulletStartPos: {fileID: 4781081830088815303}
+--- !u!114 &52368431707081983
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 9110048099626921029}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a8b0997c841f6744999af064132f614b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!95 &349551690098673573
 Animator:
   serializedVersion: 3
