+++ Assets/Scripts/Boss.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b0eed72279077447aa7c7a16840aad1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
