+++ Assets/Scripts/Boss/BossAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 13b59b0c6d24e174ca3eb084293b6d4c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
