+++ Assets/Scripts/Boss/BossController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 00a9fb7e80fe93f42b426b2846c2f8db
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
