+++ Assets/Scripts/Boss/BossMovement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3096a70c92a023e4788f380e176df0a7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
