+++ Assets/Scripts/BossTrigger.cs
@@ -7,18 +7,6 @@ public class BossTrigger : MonoBehaviour
     [SerializeField] private Door bossDoor;
     [SerializeField] private GameObject healthLineObj;
 
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-
     private void OnTriggerEnter(Collider other)
     {
         if (other.CompareTag("Player"))
