+++ Assets/Scripts/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f7b69526fe3759f438d35838f525107a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
