+++ Assets/Scripts/Enemy/EnemyAttack.cs
@@ -1,0 +1,16 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemyAttack : MonoBehaviour, IAttack
+{
+    [SerializeField] private GameObject bulletPref;
+    [SerializeField] private Transform bulletStartPos;
+
+    public void CreateBullet()
+    {
+        //Instantiate(_bulletPref, _bulletStartPos.position, transform.rotation);
+        var rBody = Instantiate(bulletPref, bulletStartPos.position, Quaternion.identity).GetComponent<Rigidbody>();
+        rBody.velocity = bulletStartPos.forward * 15f;
+    }
+}
