+++ Assets/Scripts/Enemy/EnemyBehave.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 747355fd32eb1d246883d5e93c895d92
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
