+++ Assets/Scripts/Interfaces.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1adeb46cdeb11d541b0f4f24fc9fbf79
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
