+++ Assets/Scripts/Interfaces/IAttack.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface IAttack
+{
+    public abstract void CreateBullet();
+}
