+++ Assets/Scripts/Interfaces/IAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 729adff7df6a0f7429672c6f7df63b2a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
