+++ Assets/Scripts/Interfaces/IBombAttack.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8efd326921134ad418f79591b7d9ca80
+guid: 37345ff74419f9342a8610519d020b64
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
