+++ Assets/Scripts/Interfaces/ITakeDamage.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface ITakeDamage
+{
+    public abstract void TakeDamage(int damage);
+}
