+++ Assets/Scripts/Interfaces/ITakeDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f10ecf58b3d9fd449c6f27b2f1b5448
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
