+++ Assets/Scripts/Interfaces/ITakeExplosionDamage.cs
@@ -1,0 +1,8 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public interface ITakeExplosionDamage
+{
+    public void DeadByExplosion();
+}
