+++ Assets/Scripts/Interfaces/ITakeExplosionDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46c1d788e3559bb4784a05d0a408c8f1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
