+++ Assets/Scripts/LiveObj.cs
@@ -8,7 +8,4 @@ public abstract class LiveObj : MonoBehaviour
     protected int currentHP;
 
     public bool IsAlive { get; protected set; }
-
-    public abstract void TakeDamage(int damage);
-    public abstract void DeadByExplosion();
 }
