+++ Assets/Scripts/Player/PlayerAttacking.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6f732432a9accaa49b8a6b130de56817
+guid: fee152a30c4453541a8ded30cdf50ca4
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
