+++ Assets/Scripts/Player/PlayerJumping.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 64828354aa7fe794f8f6b6060dbdbbeb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
